requirements-test.txt
pre-commit
pytest
pytest-cov
translate-toolkit
twine
